fixes unconfirmed tx and organizes scan structure
