+++ src-tauri/src/services/record_handling.rs
@@ -1,6 +1,6 @@
 pub mod decrypt_transition;
 pub mod records;
-pub mod records_new;
+pub mod scan_utils;
 pub mod sync;
 pub mod transfer;
 pub mod utils;
