+++ src/services/util/migrate_data.ts
@@ -1,0 +1,5 @@
+import {invoke} from '@tauri-apps/api/core';
+
+export async function updateData() {
+	return invoke('migrate_encrypted_data');
+}
